v0.15
*NEW*

- First iteration of dynamic player camera
- Black collision boxes added into map to restrict player view from the
outside of the map
- Smaller map to test
- New map coordinates system that allows the GUI to work in any sort of
resolution or view change
- New resource handler classes to allow for easier resource management

*MISC AND FIXES*
-Lots of GUI bug fixes
-Resolution scaler bug fixes
-Reformatted map data to look cleaner
- Work started on player settings parser

+++ Render/TileMap.hpp
@@ -57,7 +57,7 @@ public:
         return false;
     }
 
-    TileEntityRef::Ptr getTileMouseOver(sf::Vector2i& mouseCoords)
+    TileEntityRef::Ptr getTileMouseOver(sf::Vector2f& mouseCoords)
     {
         for (auto& iter : m_tiles)
         {
